requirements.txt
Pillow
tqdm
numpy
json
torch
torchvision
plotly
nltk
spacy
csv
scikit-learn
transformers
faiss
hdbscan